Refactor: 클래스 모듈 이름 변경
