+++ frontend/src/utils/index.ts
@@ -1,4 +1,4 @@
-import Fetcher from 'src/utils/fetcher';
+import Fetcher from 'src/utils/Fetcher';
 
 // eslint-disable-next-line import/prefer-default-export
 export { Fetcher };
